j773: Create a seperate utils folder to contain utility code
The utils directory contains modules that was in the steps directory but was not
an active step in the pipeline, i.e. its code was a utility to help the steps
run. The logical fix for this is to move iofuns, config_tools and the old utils
module to this new directory. Module imports have been amended to reflect this
change and the move mirrored in the test directory. Also redundant util tests
for previously deleted functions have been removed.