j513: Update path specification in main and log output location
Previously in main.py we specified the config_dir, however with logging now
taking place it seems more prudent to save the logs within the cprices repo
(untracked due to gitignore changes) for simple traceability. The change of
the config_dir to root_dir variable (with changes where appropriate in the
config loaders) means we can pass the root dir to the log_config and also
makes the code more bullet proof when being run on different systems (e.g.
Prod/DevTest).