PADSP-1026 add multi_item_sep_char to dev_config
multi_item_sep_char defines the string used as the seperator for multi item
webscraped datasets and is used in classification and aggregation. By setting
as a variable in the dev_config it is no longer mysteriously hidden away in the
code (and so open to being forgotten) and also allows more complex seperators
to be used if they intefere with the data (e.g. previously we were using a hard
coded '_' which meant that multi item datasets couldn't have a '_' in their
name!). For the time being multi_item_sep_char='_#_' hopefully a less common
string.