PADSP-778 add read of lc data in iofuncs
now possible to read the historic data stored in hdfs straight into the core
pipeline when conventional data source is specified (as in scenario_3.yaml).
Currently the pipeline is hardcoded to only read in the historic lc data (as
this is all we have and it's unclear what form the future data will be in).