PADSP-1011: Added chunking tests and fixed bugs
One bug of note was that when using the .query method an error was thrown
because the test data column name 'class' is a Python keyword. The error said
this was not allowed so changed the method to avoid the bug in the future.

Didn't use classes to group tests because I don't think it was needed.