PADSP-1398: Remove weights from config data
The weights will be needed to be read in as a separate file going forward
since they will be a time series. They may also be precise numbers so manually
inputting them in the config is not viable. The config has been changed to
remove the weights and just list the suppliers and items to include (for
retailer weights).

The iofuns.py file has been changed to work with the new configuration, as has
config_tools.py.