PADSP-1389: Change option date_trim_strategy to week_selection
This tidied things up a bit and is a bit more simple and straightforward
for the user to use. It also provides as much flexibility as they need,
as the user just specifies the weeks to keep rather than using some
preset settings. The validation still stops them putting any integers
higher than 4 in their selection though.