PADSP-1398: Remove drop_retailers option
Add the weights_dir to load in webscraped retailer weights directly for
use in the aggregation stage. They were being used also to drop some
rows in the preprocessing but as far as I could see this seemed
unnecessary - because the data to include is chosen using the scenario
files and anything that has a zero weight won't contribute anything
when aggregated.