PADSP-1371: Outlier detection refactor
There is now a single entry point to the outlier_detection methods
called `classify_outlier`. It takes a single Spark df as opposed to the
dict of Spark DFs that were done previously. This is in line with the
goals of the alt sources pipeline refactor, and is better practice.

The calculating of the fences is the main differentiation between
outlier classification methods, with the rest of the logic being mostly
the same. Some require some summary stats to be computed. These must be
computed beforehand since they use Window functions which don't work as
part of a long lineage with a WHERE clause in it.

The original functionality has been retained through on option switch in
the main function: "classify_only" - when if True it doesn't drop the
outliers, but creates a new column to identify them. It means the
function needs to be called twice currently in pipeline.py, but I don't
think it's much of a problem.

I've taken some of the same API decision as I have with the index_methods
refactor since I consider outlier detection methods to be good
candidates to lift into precon. This included using some of the same
Spark helper functions and I've just duplicated them at the higher
precedence level currently, rather than importing from index_methods
(since index_methods will eventually be subsumed by precon). I've made
my first UDF! Basically just broadcasting the np.diff function across a
column. I'm not sure of it's impact on performance so needs to be kept
an eye on.