PADSP-1367: Further refactor on ws preprocessing and tests
Where preprocessing steps are using well documented methods and we are
not overly concerned with scrutinising any edge cases, then it makes
sense to just put the methods straight in preprocessing.main func. But
where we want to inspect a bit further and isolate some tests, then we
break out a separate function.

There isn't really any hard rules around this and a lot is open to
interpretation so this may be a bit overbearing I admit!