PADSP-1416: Split validation and config into separate files
Decided to add the config and validation modules at the top level. The
configs now work by inheriting from a config base class, the result is a
much cleaner config definition which does the same job.

The validation has been standardised across the board, and split into
each of the sections that are in the config file. There might be some
reduction of DRY in terms of functionising some of the validation steps
but it doesn't have much priority.

Thinking about adding a feature which will catch all the errors in the
scenario files and reporting them all when raising.