PADSP-1416: Break up run function in main.py
There is now a run_scenario function that runs the separate pipelines
for each scenario. A separate function for run_local_colelction was
created since there were a fex extra steps with the data loading and
chunking. For scanner and web_scraped, a new module has been created in
each sub-pipeline called loaders.py and the data is now actually loaded
as part of each pipeline so they are more self-contained and things look
clearer in the `run_scenario` function.

When more is known about the interplay between the histroic and ongoing
datasets in the local collection, then perhaps in the future the loading
can be done as part of the local collection pipeline. In truth, there
probably needs to be another module which does most of the actually
stats processing in local collection, then a separate pipeline file that
loads the data also. It doesn't really belong in the current
pipeline.main() func.