PADSP-1416: Improve saving to HDFS code in main.py
I've improved the concat function in spark_helpers, added a helper
function to inverse dictionary nesting (using flatten_dict) and added a
new function called combine_scenario_df_outputs to improve the way the
outputs are saved and do away with a lot of unnecessary helper
functions.

In pipeline.py I convert the two pandas DataFrame outputs to spark
before the main function returns, so all the outputs coming out are
consistent type. This simplifies things later on and is explicit.