PADSP-1580: Add new factory func for transformer decorators
An example of a transformer decorator is to_spark_col as it takes
function arguments and transforms them before they are used. Another one
that I wanted to have was to_list, to convert certain function arguments
to list where the expected type is Union[str, Sequence[str]] (strings
behave like a list so it gets a little complicated - need to wrap in
[]).

So rather than copy and paste the code, I created a factory function
that allows us to create these transformer decorators on the fly using
only the simple transformation function. While doing so, I moved a few
functions around into different modules and made sure to fix all the
imports and use of certain functions.