PADSP-1580: Working calculate_indices with new options format
The new options format has been developed off the back of requirements
from the research and index_methods teams. What we now have is nested
options for each of the multilateral methods, so that the options there
are more explicit. I've also adapted the calculate_indices code to take
the parameters `bilateral_index_options` and
`multilateral_index_options`. `bilateral_index_options` is simply a
mapping of all the function params for the bilateral_pipeline.
`multilateral_index_options` is similar, except there as an additional
level of nesting so that the options can be supplied for each different
multilateral method.

There are still quite a few design choices to make:
* should the output have a MultiIndex or not?