PADSP-1579: Split config into separate pipelines
The config folder now has sub-directories for each pipeline. This works
because the pipelines are only intended to be run separately, as they
will now be combined in the aggregation pipeline which comes subsequent
to this pipeline.

Also created a different dev_config for each of the pipelines, so that
the various columns can be set. It's expected that there will be
different strata depending on whether it's scanner or web scraped so
this helps keep them separate.

Also put in some initial work on the validation, so that there are
separate validation functions for each pipeline - since we expect
different sections for the scenario files depending on the type of
pipeline being run.

Still needs work...