PADSP-1579: Initial attempt to make the validation more DRY
Trying to dynamically generate errors as a lot of them take similar
forms. The next step I want to do is to try and collect all the errors
first before raising anything so that the user doesn't have to go back
and run multiple times before finding all the errors.