PADSP-1603: Implemented split scenario configs
Required making changes to the `main` module and the pipeline modules
for web_scraped and scanner. In main, each type of scenario is looped
over and then the dataframes joined at the end. This allows us to modify
the separate dev_configs, without changing the dev_config for other
runs.

The local collection has currently been left out of the changes - the
implementation of this should come next.

Now the pipelines are properly split which will make the maintenance
much easier when trying to change the config for one of either the
webscraped or scanner pipeline, rather than doing convoluted processes
to get around the fact they are combined but require different settings.