PADSP-1567: Columns added/removed dependent on scenario config
The discount_col is now only added if remove_discounts=True in the
scenario config.

If the use_unit_prices option is False, then standardised_size and
weight_type_ons are removed from the data cols.

Both these changes are aimed at minimising the size of the dataframe to
only columns needed, for optimisation purposes.